web/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-countries==7.4.2
django-crispy-forms==1.14.0
gunicorn==21.2.0
packaging==23.1
Pillow==9.3.0
psycopg2-binary==2.9.7
sqlparse==0.4.3
typing_extensions==4.7.1
